Introduced the TemplatingConfigurationManager which sole purpose is to -independent from any other configuration management- read, load and process templating related configuration.
Added a ConfigUtils class which provides common useful methods in terms
of configuration management

Added an example / sample of templating.config and an example / sample
of a requirment template file. Those files are in the resources
directory to have them accisible in the dev einvorment (IntelliJ Idea)
The configuration management respects whether the application is packed
into a jar or is executed from commandline.
Introduced the TemplatingConfigurationManager which sole purpose is to
-independent from any other configuration management- read, load and
process templating related configuration.

Added a ConfigUtils class which provides common useful methods in terms
of configuration management

Added an example / sample of templating.config and an example / sample
of a requirment template file. Those files are in the resources
directory to have them accisible in the dev einvorment (IntelliJ Idea)
The configuration management respects whether the application is packed
into a jar or is executed from commandline.